+++ macos/Onit/PanelStateManager/Tethered/PanelStateTetheredManager.swift
@@ -166,7 +166,7 @@ class PanelStateTetheredManager: PanelStateBaseManager, ObservableObject {
                 var action = action
                 
                 if state.panelWasHidden {
-                    state.tempShowPanel()
+                    tempShowPanel(state: state)
                 }
                 
                 // TODO: KNA - We need to store the frame used to calculate the tetheredButtonYPosition and apply a diff
@@ -182,7 +182,7 @@ class PanelStateTetheredManager: PanelStateBaseManager, ObservableObject {
                 hideTetherWindow()
 
                 if state.currentAnimationTask == nil {
-                    state.repositionPanel(action: action)
+                    showPanel(for: state, action: action)
                 }
             } else {
                 // Panel closed
@@ -192,7 +192,7 @@ class PanelStateTetheredManager: PanelStateBaseManager, ObservableObject {
         } else {
             // If it's hidden, we want to hide the tether window and potentially animate out the panel.
             if (state.panelOpened && !state.panelWasHidden) {
-                state.tempHidePanel()
+                tempHidePanel(state: state)
             }
             hideTetherWindow()
         }
