+++ macos/Onit/UI/Panels/State/OnitPanelState+ScreenMode.swift
@@ -5,6 +5,8 @@ import Foundation
 import AppKit
 
 extension OnitPanelState {
+    
+    // TODO: KNA - Refacto: Should be removed at the end
     var isScreenMode: Bool {
         return trackedWindow == nil && trackedScreen != nil
     }
